Finish CRUD Pokemon

+++ server/Repository/CategoryRepository.cs
@@ -49,9 +49,7 @@ namespace server.Repository
                 var category = await _context.Categories.FindAsync(categoryId);
 
                 if (category == null)
-                {
                     throw new Exception("Category not found!");
-                }
 
                 _mapper.Map(categoryDto, category);
                 _context.Categories.Update(category);
@@ -72,9 +70,7 @@ namespace server.Repository
                 var category = await _context.Categories.FindAsync(categoryId);
 
                 if (category == null)
-                {
                     throw new Exception("Category not found!");
-                }
 
                 category.Hidden = !category.Hidden;
                 _context.Categories.Update(category);
@@ -90,7 +86,12 @@ namespace server.Repository
         {
             try
             {
+                var checkValid = _context.Categories.Where(c => categoryIds.Contains(c.Id)).Count();
+                if (checkValid != categoryIds.Length)
+                    throw new Exception("One or more categories not found!");
+
                 var categories = await _context.Categories.Where(c => categoryIds.Contains(c.Id)).ToListAsync();
+
                 foreach (var category in categories)
                 {
                     category.Hidden = !category.Hidden;
@@ -109,13 +110,15 @@ namespace server.Repository
             try
             {
                 var category = await _context.Categories.FindAsync(categoryId);
-
                 if (category == null)
-                {
                     throw new Exception("Category not found!");
-                }
-
                 _context.Categories.Remove(category);
+
+                var pokemonsCategory = await _context.PokemonCategories.Where(c => c.CategoryId == categoryId).ToListAsync();
+                if (pokemonsCategory == null)
+                    throw new Exception("Pokemon category not found!");
+                _context.PokemonCategories.RemoveRange(pokemonsCategory);
+
                 await _context.SaveChangesAsync();
             }
             catch (Exception ex)
@@ -128,8 +131,16 @@ namespace server.Repository
         {
             try
             {
+                var checkValid = _context.Categories.Where(c => categoryIds.Contains(c.Id)).Count();
+                if (checkValid != categoryIds.Length)
+                    throw new Exception("One or more categories not found!");
+
                 var categories = await _context.Categories.Where(c => categoryIds.Contains(c.Id)).ToListAsync();
                 _context.Categories.RemoveRange(categories);
+
+                var pokemonsCategory = await _context.PokemonCategories.Where(c => categoryIds.Contains(c.CategoryId)).ToListAsync();
+                _context.PokemonCategories.RemoveRange(pokemonsCategory);
+
                 await _context.SaveChangesAsync();
             }
             catch (Exception ex)
